(refactor) CommandLineRunner 클래스 분리
